% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prep_fqgam_data}
\alias{prep_fqgam_data}
\title{Prepare Data for Functional Quantile GAM}
\usage{
prep_fqgam_data(y, id, time, Xobs, s_range = c(0, 1))
}
\arguments{
\item{y}{Numeric vector. Response variable.}

\item{id}{Factor or character vector. Subject identifiers.}

\item{time}{Numeric vector. Longitudinal time stamps.}

\item{Xobs}{Matrix. Functional covariate observations where rows are
observations and columns are functional grid points.}

\item{s_range}{Numeric vector of length 2. Range of the functional domain.
Default is \code{c(0, 1)}.}
}
\value{
A data.frame suitable for use with \code{qgam}, containing:
\describe{
\item{y}{Response variable}
\item{id}{Subject identifier (as factor)}
\item{time}{Longitudinal time}
\item{Xobs}{Functional covariate matrix}
\item{sMat}{Matrix of functional domain coordinates}
\item{tMat}{Matrix of longitudinal time coordinates}
}
}
\description{
Helper function to prepare longitudinal functional data for use with
\code{\link[qgam]{qgam}}.
}
\details{
This function creates the required matrices for the tensor product smooth
representation of the functional coefficient. The model uses:
\itemize{
\item \code{sMat}: Matrix of functional domain coordinates, where each row
corresponds to an observation and contains the functional grid points
\item \code{tMat}: Matrix of longitudinal time coordinates, where each row
contains the same time value repeated S times
}

The functional covariate enters the model formula as:
\code{te(sMat, tMat, by = Xobs, bs = c("cr", "cr"), k = c(k_s, k_t))}

where the \code{by} argument allows for the integration approximation.
}
\examples{
\dontrun{
library(refund)
data(DTI)
myData <- DTI[complete.cases(DTI$cca), ]
myData <- subset(myData, Nscans > 1)

myDataFit <- prep_fqgam_data(
  y = myData$pasat,
  id = myData$ID,
  time = myData$visit.time,
  Xobs = myData$cca
)
}

}
