% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.fqgam_bias_adj}
\alias{print.fqgam_bias_adj}
\title{Print Method for fqgam_bias_adj Objects}
\usage{
\method{print}{fqgam_bias_adj}(x, ...)
}
\arguments{
\item{x}{An object of class \code{fqgam_bias_adj}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print Method for fqgam_bias_adj Objects
}
