% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.fqgam_boot}
\alias{print.fqgam_boot}
\title{Print Method for fqgam_boot Objects}
\usage{
\method{print}{fqgam_boot}(x, ...)
}
\arguments{
\item{x}{An object of class \code{fqgam_boot}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print Method for fqgam_boot Objects
}
