% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.fqgam_ci}
\alias{print.fqgam_ci}
\title{Print Method for fqgam_ci Objects}
\usage{
\method{print}{fqgam_ci}(x, ...)
}
\arguments{
\item{x}{An object of class \code{fqgam_ci}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print Method for fqgam_ci Objects
}
